refactor: split GameScene.swift and introduce BallCoordinator (#119)
- Extract setup (didMove, setupCamera, configurePhysics, setupBackground,
  setupNodes) → GameScene+Setup.swift
- Extract input handling (handle, movePaddle, touches, mouse/keyboard) →
  GameScene+Input.swift
- Restore GameScene+Boss.swift (setupBossCoordinator, applyBossTickResult,
  handleBossLifeLoss)
- Introduce BallCoordinator: owns balls[], cullFallen, spawnExtra,
  removeExtras, applyEffects — absorbs the multi-ball and power-up
  effect overloads that were in GameScene
- Fix pre-existing macOS build error: replace Optional.map pattern with
  if let for @MainActor BossCoordinator calls

GameScene.swift shrinks from 480 → 191 lines.
